+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.5](https://github.com/alxhill/superline/compare/v0.5.4...v0.5.5) - 2026-06-26
+
+### Added
+
+- *(config)* sync default config with recommended setup ([#25](https://github.com/alxhill/superline/pull/25))
+
+### Fixed
+
+- *(cwd)* show current dir under Git Bash on Windows ([#27](https://github.com/alxhill/superline/pull/27))
+
 ### Fixed
 
 - *(cwd)* show the current directory under Git Bash on Windows. The native